[modify] rename files, uncapitalize files' name
