+++ patterns/Architectual/Model_View_ViewModel/index.html
@@ -15,7 +15,7 @@
 	</div>
 
   <script type="module">
-		import Mvvm from './main.js'
+		import Mvvm from './js/main.js'
 	</script>
 </body>
 </html>
