+++ patterns/Architectual/Model_View_ViewModel/js/model.js
