+++ patterns/Architectual/Model_View_ViewModel/js/viewModel/watcher.js
